bin/adb-key: make them up case bin/after-co-ln-s.sh: do not do it on mac

+++ bin/after-co-ln-s.sh
@@ -27,7 +27,7 @@ done
 
 mkdir -p ~/.logs
 
-if which sudo && test ! -e ~/.logs/offline-is-unstable; then
+if which sudo && test $(uname)  = Linux -a ! -e ~/.logs/offline-is-unstable; then
     sudo apt-get install -t unstable offlineimap
     touch ~/.logs/offline-is-unstable
 fi
